Fix Makefile ^M character (grepping for bad double newlines) Remove trailing whitespace so we can compile everything First attempt to format y-axis

+++ Makefile
@@ -121,8 +121,7 @@ protovis-r3.3.js: $(JS_FILES)
 
 %-d3.3.js: Makefile
 	grep '	' -Hn $(filter %.js,$^) && echo "ERROR: tab" && exit 1 || true
-	grep '
-' -Hn $(filter %.js,$^) && echo "ERROR: dos newline" && exit 1 || true
+	grep '' -Hn $(filter %.js,$^) && echo "ERROR: dos newline" && exit 1 || true
 	grep ' $$' -Hn $(filter %.js,$^) && echo "ERROR: trailing space" && exit 1 || true
 	rm -f $@
 	echo "// $(shell git rev-parse HEAD)" >> $@
